+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+#include "generated_plugin_registrant.h"
+
+#include <window_utils.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  WindowUtilsRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("WindowUtils"));
+}