+++ lib/window_utils.dart
@@ -14,16 +14,76 @@ class WindowUtils {
     return _channel.invokeMethod<bool>('hideTitleBar');
   }
 
-  static Future<bool> setMoveable(bool canMove) {
-    return _channel.invokeMethod<bool>('setMovable', {"canMove": canMove});
+  static Future<bool> closeWindow() {
+    return _channel.invokeMethod<bool>('closeWindow');
+  }
+
+  /// [Windows] Only
+  static Future<bool> minWindow() {
+    return _channel.invokeMethod<bool>('minWindow');
+  }
+
+  /// [Windows] Only
+  static Future<bool> maxWindow() {
+    return _channel.invokeMethod<bool>('maxWindow');
+  }
+
+  static Future<bool> centerWindow() {
+    return _channel.invokeMethod<bool>('centerWindow');
+  }
+
+  static Future<bool> setPosition(Offset offset) {
+    return _channel.invokeMethod<bool>('setPosition', {
+      "x": offset.dx,
+      "y": offset.dy,
+    });
+  }
+
+  static Future<bool> setSize(Size size) {
+    return _channel.invokeMethod<bool>('setSize', {
+      "width": size.width,
+      "height": size.height,
+    });
   }
 
   static Future<bool> startDrag() {
     return _channel.invokeMethod<bool>('startDrag');
   }
 
+  /// [Windows] Only
+  static Future<bool> startResize(DragPosition position) {
+    return _channel.invokeMethod<bool>(
+      'startResize',
+      {
+        "top": position == DragPosition.top ||
+            position == DragPosition.topLeft ||
+            position == DragPosition.topRight,
+        "bottom": position == DragPosition.bottom ||
+            position == DragPosition.bottomLeft ||
+            position == DragPosition.bottomRight,
+        "right": position == DragPosition.right ||
+            position == DragPosition.topRight ||
+            position == DragPosition.bottomRight,
+        "left": position == DragPosition.left ||
+            position == DragPosition.topLeft ||
+            position == DragPosition.bottomLeft,
+      },
+    );
+  }
+
+  static Future<bool> windowTitleDoubleTap() {
+    return _channel.invokeMethod<bool>('windowTitleDoubleTap');
+  }
+
+  /// [MacOS] Only
   static Future<int> childWindowsCount() {
     return _channel.invokeMethod<int>('childWindowsCount');
+  }
+
+  /// Size of Screen that the current window is inside
+  static Future<Size> getScreenSize() async {
+    final _data = await _channel.invokeMethod<Map>('getScreenSize');
+    return Size(_data['width'] as double, _data['height'] as double);
   }
 
   static Future<Size> getWindowSize() async {
@@ -35,4 +95,15 @@ class WindowUtils {
     final _data = await _channel.invokeMethod<Map>('getWindowOffset');
     return Offset(_data['offsetX'] as double, _data['offsetY'] as double);
   }
+}
+
+enum DragPosition {
+  top,
+  left,
+  right,
+  bottom,
+  topLeft,
+  bottomLeft,
+  topRight,
+  bottomRight
 }