+++ CHANGELOG.md
@@ -1,3 +1,25 @@
-## 0.0.1
+# Changes
 
-* TODO: Describe initial release.
+## 1.0.1
+
+- Removed: `setCanMove`
+- Added: `minWindow` (Windows)
+- Added: `maxWindow` (Windows)
+- Added: `getScreenSize` (MacOS, Windows)
+- Added: `setPosition` (MacOS, Windows)
+- Added: `setSize` (MacOS, Windows)
+- Added: `closeWindow` (MacOS, Windows)
+- Added: `centerWindow` (MacOS, Windows)
+- Updated: `getWindowSize` (Windows)
+- Updated: `getWindowOffset` (Windows)
+- Added: `windowTitleDoubleTap` (MacOS, Windows)
+
+## 1.0.0
+
+- Added: `showTitleBar` (MacOS)
+- Added: `hideTitleBar` (MacOS)
+- Added: `setCanMove` (MacOS)
+- Added: `startDrag` (MacOS)
+- Added: `childWindowsCount` (MacOS)
+- Added: `getWindowSize` (MacOS)
+- Added: `getWindowOffset` (MacOS)